refactor: rename dataset component files to kebab-case for consistency
